+++ grails-app/services/io/xh/hoist/monitor/MonitorService.groovy
@@ -41,10 +41,9 @@ class MonitorService extends BaseService {
 
     // Shared state for all servers to read - gathered by primary from all instances.
     // Map of monitor code to aggregated (cross-instance) results.
-    private CachedValue<Map<String, AggregateMonitorResult>> _results = new CachedValue<>(
+    private CachedValue<Map<String, AggregateMonitorResult>> _results = createCachedValue(
         name: 'results',
-        replicate: true,
-        svc: this
+        replicate: true
     )
 
     private Timer timer
@@ -52,7 +51,8 @@ class MonitorService extends BaseService {
     void init() {
         timer = createTimer(
             name: 'runMonitors',
-            interval: { monitorInterval },
+            runFn: this.&runMonitors,
+            interval: {monitorInterval},
             delay: startupDelay,
             primaryOnly: true
         )
@@ -87,7 +87,7 @@ class MonitorService extends BaseService {
     //------------------
     // Implementation
     //------------------
-    private void onTimer() {
+    private void runMonitors() {
         // Gather per-instance results from across the cluster
         Map<String, List<MonitorResult>> newChecks = clusterService
             .submitToAllInstances(new RunAllMonitorsTask())