+++ src/main/groovy/io/xh/hoist/role/provided/DefaultRoleService.groovy
@@ -81,10 +81,9 @@ class DefaultRoleService extends BaseRoleService {
     DefaultRoleUpdateService defaultRoleUpdateService
 
     private Timer timer
-    protected CachedValue<Map<String, Set<String>>> _allRoleAssignments = new CachedValue(
+    protected CachedValue<Map<String, Set<String>>> _allRoleAssignments = createCachedValue(
         name: 'roleAssignments',
         replicate: true,
-        svc: this,
         onChange: {
             _roleAssignmentsByUser = new ConcurrentHashMap()
         }
@@ -103,8 +102,8 @@ class DefaultRoleService extends BaseRoleService {
 
         timer = createTimer(
             name: 'refreshRoles',
-            interval: { config.refreshIntervalSecs as int * SECONDS },
             runFn: this.&refreshRoleAssignments,
+            interval: { config.refreshIntervalSecs as int * SECONDS },
             runImmediatelyAndBlock: true,
             primaryOnly: true
         )